polymarket_backtest/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
urllib3>=2.0.0
python-dotenv>=1.0.0
